live2d: Create a second copy of each mesh to use as a mask.
This is because the first copy was getting its shaders changed,
which would cause a problem when that copy was being used as a
mask.

This would cause a problem when a live2d layer both used a mask
and was used as a mask.

Per https://lemmasoft.renai.us/forums/viewtopic.php?f=8&t=66455