Remove the changes to polar coordinates, start over.
This removes the ability of radius to be given as a float, which
doesn't make a ton of sense. Is the radius measured compared to
width, height, or something else? Now it's back to being pixels,
which is was in 8.0.

This does make around work better when it is a float, using the
available area.

Polar coordinates now always produce absolutes for xpos and ypos,
allowing radius and around to be in different types.