Allow loading premultiplied textures.
This is to support the changes to text that produced premultiplied
surfaces, rather than non-premultiplied surfaces. This was necessary
to display emoji properly, and also speeds up the case of normal
text rendering, at least on gl2.

This also removes some old paths that worked with GL_BGRA, which
we aren't using.