Compat out
The idea is to emulate the former behavior, in which the return type of position interpolations was the type of b, so that the position type is never instanciated.
That to keep it from ending up in existing games' renpy.get_placement, which expect only (real) numbers.
I tried to minimize the computation burden, too.