Move away from Formatter for string interpolation
As we were already overriding the C parser, the only advantage we were
taking advantage of was the C field name parser. We continue to use this
but by moving away from the string.Formatter template we can ignore
a bunch of things that we aren't interested in (such as which fields
from the supplied scope were used).