Reimplement apply in a much better way
The default values are now evaluated lazily, only when needed, which is both faster and closer to the previous behavior.
The implementation of ignore_errors is not as good, but that should get improved soon.
Performances of the apply call on autorange goes from 50k-.30 to 100k-.32, which is almost a 2x speedup
The legacy apply call is at 100k-.23, we're dragging 40% slower but closing