text: Introduce split ignore.
Text that's marked with SPLIT_IGNORE (the unicode
CL and CP classes) is text that isn't considered for the
purpose of the linebreaking algorithm. These characters tend to
be CJK punctuation and close parenthesis.

SPLIT_IGNORE text doesn't have it's size taken into account when at
the end of a line. This keeps the layout the same visual width, rather
than breaking a character and it's punctuation on the next line.

Fixes #3356, except for ー , though renpy.language_tailor could be
used to change the break class of that character if desired.