Optimize is_pixel_opaque to consider Render transparency.
This is a performance improvement to focus_mask, especially in
the case where where the bounding of non-transparenty pixels is
small relative to the size of the imagebutton. In that case, the
GPU is only invoked when inside the bounding box.

This prevents expensive round trips to the GPU when we're sure
those operations will be unproductive.

Per #3612.