Add the ability to defer the loading of translation scripts.
This speeds up the start of games with large numbers of
translations, by deferring the loading of a game's
translation scripts until the language is selected.

The downside is that these scripts are loaded out of order, but
that seems like a small price to pay for the speed boost that
can be provided.