Disable fullscreen on Safari.
This is due to several problems that don't seem to occur on other
browsers. While the visibility problems were patched away, once
we went fullscreen, it was impossible to show anything else, like
the web_input textbox.

After a weekend working on it, no luck - so Fullscreen is disabled.

Fixes #5038.