Treat string interpolations as Python expressions (#5036)
* Move away from Formatter for string interpolation

As we were already overriding the C parser, the only advantage we were
taking advantage of was the C field name parser. We continue to use this
but by moving away from the string.Formatter template we can ignore
a bunch of things that we aren't interested in (such as which fields
from the supplied scope were used).

* Whitespace

* Manage conversion flags with sets

* Update interpolation parser to support expressions

This change allows strings and nested brackets, braces, and parens to be
within an interpolation without breaking out. It also allows conversions
to be specified before the format (in line with Python), with graceful
fallback to ensure existing interpolations continue to work as intended.

* Treat interpolation as Python expressions

* Add support for debug interpolations

This apes fstring functionality where a trailing = symbol is used to
denote that the expression part of an interpolation should be included
with the output and, when no conversion or format are specified, imply
the result should be run through repr.

* Update interpolation documentation for expressions

* Add interpolation expressions to changelog