Only predict_show a screen by name when not transient.
This fixes #4999 while trying to keep #4834 fixed. There is a
very minor behavior changes here:

When a screen has been called, renpy.showing will not return
true for a screen's name if the screen has a tag.

This seems fine, because renpy.showing is only documented to work
with tags, not names - working with screens by name has been the
job of renpy.get_screen. So it's only a change to undocumented
behavior.