rb: Ensure that all checkpoints are tracked
Once the hard_rollback_limit is exceeded, hard checkpoints are added to
the number of blocked checkpoints, ensuring that when totalled limit and
block counts total the number of hard checkpoints currently in the log.

Now that the entire log is counted, it makes it safe to count hard
checkpoints out of the log, first from the blocked count, and then from
the limit count, never dropping below zero.