pos: Drop ATL *_or_none functions
The rationale for this is that they didn't serve their intended purpose,
and somewhat misleading. They only called by interpolation, and in such
a way that the argument could never be None.

In 8.2 a special case was added to spline interpolation which used one
of these functions directly, however the additional None handling failed
to make the result any more usable. In order to address this in a more
useful way, and align with non-spline interpolation, None values will
now be treated as zero.