Allow ren'py 8 to load .rpyc files using old-style screens.
Ren'py 8 still has all the code needed to support these files, but
as the ast module of python has changed between 2.7 and 3.9 it becamse
impossible to even load these files as the unpickling process would error
due to unknown classes. This means nearly any ren'py 6.17 game cannot be ported
over without a recompile.

This commit fixes this by intercepting the construction of these classes
using pre-existing mechanisms in renpy.compat.pickle. When not needed this
adds only a single equality check to the critical path of unpickling files.