Change how window auto works.
The window hide/show statements no longer disable the window auto
flag, with new statements (window auto True, window auto False)
taking care of that.

The thinking behind this is that window hide is useful in it's own
right, but significantly less useful if it requires a matching
window auto each time.

For example, one might want to do:

     e "Dialogue"
     window hide
     show clouds with dissolve
     e "Dialogue"

and have that work.