text shader: Create meshes.
This moves the conversion of positions to texture coordinates
to glyph_coordinates, where it's faster.

This also stores information about RTL and duration into
each glyph, which is then used to figure out the duration
of each glyph.

Finally, the mesh is created when it doesn't exist already,
and stored in the layout.