Changes to box_align and box_justify.
This moves alignment and justification to a second pass, when the
size of each placement is fully known. This seems better than trying
to place each with partial information and then correcting it.

As a change, box_align and box_justify no longer imply xfill or
yfill. This follows how Text works, where justify and align
occur inside a small text size.