input: Improve rendering of caret
This change ensures that there is always space in the input render for
the caret, while retaining the current behaviour where shifting the
caret left and right does not impact the layout of the text.

This allows us to go back to using a zero-width space as a sentinel in
empty inputs (to ensure their correct height) thus avoiding ambiguity
between a sentinel space and a user-provided space.