Avoid save scanning restarting the interaction.
The strategy chosen was to keep track of which slots have had
information by the rest of Ren'Py, and only restarting if one of
those slots is accessed.

This is motivated by a series of interaction restarts at choice menus,
caused by the autosaves that happen when a choice menu is reached.