Many changes to how Ren'Py deals with seen statements.
The biggest one is that once a translation has been seen, dialogue
statemments inside that translation are also marked as seen.

This change also changes the representation of seen translates and
nodes, to store a 64-bit hash of the identifier, which should take
up less memory and should be able to be pickled faster.