Make game/libs and game/mods special.
If flag files, game/libs/libs.txt and game/mods/mods.txt exist.

In both cases, the files in the directory load in unicode order by
filename, and not by full path. Files in game/libs load before other
files in game, while those in game/mods load after other files in
games.

The main reason for this is to support CDSes that are  distributed
in libraries. A CDS in game/libs/cds.rpy can be used in any file
in the main game.

Similarly, mods son't affect the parsing of the main game.

Per #6049.