New commit to add logic for folder detect
Re-factored this to instead of using JSON bundles to assign folders; it auto detects folders one level down.  Those that contain valid games are added to the launcher, those that are not are ignored.  Hard coded a version of the tutorials to do the same.  Currently have all other projects in the main folder listed, could use a folder to contain those too if desired in a future update.