Track column numbers during parsing.
How these are handled varies for PyCode and PyExpr.

PyCodes are, for now, parsed with all the indentation present in the
file. This is then passed to the parser, which pads the PyCode out
with an if True statement so it compiles.

PyExprs have the column number determined by the parser, and it's
stored in the object, to be passed to the compiler.