live2d: Use the pixel size for masks.
Previously, Ren'Py had been assuming Live2D only used coordinates
in the -1.0, 1.0 range, but that isn't the case. That worked some
of the time, and always with non-masked layers, but masks couldn't
represent coordinates outside that range.

This tended to manifest as missing eyes.

This uses the correct size for masks, fixing this problem.

Fixes #6183.