vscode: Better generation of renpy.json.
It's now possible to mark classes and functions as actions. This is done
automatically if the function inherits from Action, but if not,
:doc: file_name action is the way to go.

Actions are now marked as Action in renpy.json.

The renpy/internal distinction is back, as it's expected by the vscode
plugin in a few places.

Fixes #6165.