Avoid storing persistent data replicas in memory
Ren'Py currently holds on to loaded data long past necessary, and for
larger games especially this can result in a non-trivial amount of
memory being consumed, especially as there are typically at least three
copies of persistent: game, user, and sync.

This patch frees the persistent object in FileLocation when it is pulled
for merging into the main persistent object, while leaving the mtime
value, used to determine when to reload it, untouched.

A happy side effect of this is that the amount of operations done when
evaluating whether a merge is needed is reduced significantly.