Only load a persistent file if it'll be used
Persistent data is only merged into the main persistent object if it has
a newer mtime. We already track the mtime of the last file to be merged
in, so the scan thread can use that to determine whether it's even worth
loading a persistent file.

This cuts the number of persistent files loaded during boot, resulting
in a best case scenario of a single load, and a worst case of three. An
improvement over the four achieved in the previous commit.