Optimise language initialisation
The change_language function makes both global changes such as altering
styles, and context level changes such as changing a value in a store.
Global changes are only needed when the language in use changes, but the
context level changes need to be applied each time a context is started.

This change skips the work of updating styles, screens, and other global
concerns unless the language in use has changed, but still executes the
translate blocks to correctly set up a context.

In order to prevent a memory leak due to translated styles being applied
repeatedly, we back up and restore the current styles when not actively
regenerating them as part of a language switch proper.