python: Store simple literal values instead of re-creating them with eval.
With this change, simple literal values like integers, floats, strings,
booleans, bytes, None, and complex numbers are stored instead of bytecode,
and if found, are returned directly. This is about a 75% speedup, albeit
for something fast already.

Instead of the value, a ("literal", value) tuple is stored. This
makes it so eval("string") is unambiguous, and also allows for future
expansion.