Fix fade-in for first audio clip on Web
Initializing the gain to 0 for new channels makes sure the fade-in
starts from 0 for the first audio clip played on the channels. If
no fadein is given for it, the gain will be set to max by
start_playing() and there is no reason for the first audio clip to
have the fadeout property set in start_playing().

Fixes #6291