Change the behavior of mesh_pad when left or top padding is present.
Previously, it would offset the child in a displayable the size of
the child - which isn't sensible behavior. Now, mesh_pad will ensure
the original child is in the same place and uses the same size for
layout.

As a bonus, this means that it's possible to do:

show eileen happy:
    mesh_pad (50, 50, 50, 50)
    blur 50

And have it do the right thing.

Fixes #6223.