Raise an error when playing a video stored in game.zip
Only the 'url:' scheme is supported for video playback (i.e., video files
stored on a web server).

Supporting video playback from game.zip entries is probably possible by
reading the whole file in memory using FS.readFile(), wrapping the result
in a Blob, create an URL for it using URL.createObjectURL() and using
that url as if we extracted it from a 'url:' scheme.
But storing videos in game.zip is probably a bad practice because of the
files size and the need to read the whole file in memory before being able
to play it, so do we really want to support that?