Compensate for carriage return in scriptedit
Script files are read in universal newlines mode to better leverage
various built-in Python code processing features such as linecache.

This is great for script processing, but less ideal for script editing,
where collapsed CRLFs can introduce problems with line start and end
offsets. To compensate for this, when populating scriptedit.lines we now
read the raw file and, when necessary, accumulate the drift caused by CR
characters before using it to adjust the start and end offsets.