Testsuite all, renamed reporter methods
The "all" testsuite is now the default. It contains all top-level testsuites and testcases and executes them (unless they're marked as skipped). If "all" does not exist, a new testsuite is created and populated.

Hooks for pre/post-test methods can be used with "all" to reset the state before each test, even if a test crashes.

```
testsuite all:
    before_each:
        if not screen main_menu:
            run MainMenu(confirm=False)

    after:
        exit
```