Set `spec.has_location` to `True`
By the current condition, the modules in the `.rpa` archive are assigned `spec.has_location == False`, because of this, the module does not receive the `__file__` property. In turn, this breaks the loading of module files via `pkgutil.get_data'. I consider it necessary to set `spec.has_location` to `True' without conditions.