Avoid reporting leaks of texture 0.
If OpenGL is failing, texture number generation will fail, and the
texture number will remain 0. When the gl context is reset, these
broken textures will be incorrectly reported as leaks, which is not
helpful.

This is the likely cause of #6515, and since texture 0 isn't likely to
be used in practice, we can ignore the leak.