test: Only show "Run Testcases" if default testcase exists.
Testcases are an advanced feature that will likely only be used by
a small fraction of developer, and we don't want to have an unclickable
button up all the time for everyone else.

Since it's present for the tutorial, there will be at least some
discoverability for people who want to use it.