user defined confirm screen
it adds the option to specify a screen that should be used to `renpy.confirm() `and `Confirm()` rather than the default screen `confirm`.
it might be more convenient to have separate screens for confirmation / information purposes rather than having only the default one and modify its layout based on the message to be shown.