Restore consistency to script files I/O
This allows Python to handle line endings during both read and write
operations. The result is that, regardless of a script's line-ending
choice, Ren'Py only ever has to deal with LF line-endings, simplifying
offset calculations for edits as well as ongoing maintenance burdens.

The trade-off is that output will always reflect the platform on which
Ren'Py is running, not the line-endings of the original script, i.e. on
Windows, edited scripts will always be output using CRLF.