sdl3: Update surface to SDL3.
This moves surface handling to SDL3. In the process, portions of the
API that Ren'Py does not use have been omitted. Support for surfaces
other than RGBA are gone, as are things like chroma keys and
blit modes other than alpha blending.

(These might need to come back, but for now, we'll see.)